Renamed CollectionTypeWrappedInUnmodifiableIdiomChecker to CollectionTypeWrappedInUnmodifiableIdiomOrMadeByImmutableFactoryChecker
