+++ src/main/java/org/mutabilitydetector/checkers/hint/WrappingHintGenerator.java
@@ -25,7 +25,7 @@ import com.google.common.collect.ImmutableMultimap;
 import com.google.common.collect.Multimap;
 import org.mutabilitydetector.checkers.CollectionField;
 import org.mutabilitydetector.checkers.CollectionField.GenericType;
-import org.mutabilitydetector.checkers.CollectionTypeWrappedInUnmodifiableIdiomChecker.Configuration;
+import org.mutabilitydetector.checkers.CollectionTypeWrappedInUnmodifiableIdiomOrMadeByImmutableFactoryChecker.Configuration;
 import org.mutabilitydetector.checkers.hint.exceptions.WrappingHintGenerationException;
 import org.mutabilitydetector.checkers.info.CopyMethod;
 import org.mutabilitydetector.locations.ClassNameConverter;
